modified:   data/index.html 	modified:   lib/components/servo/ServoMotor.cpp 	modified:   lib/components/servo/ServoMotor.h 	modified:   lib/logic/explorer/Explorer.cpp 	modified:   lib/logic/explorer/Explorer.h 	modified:   lib/logic/navigator/Navigator.h 	modified:   lib/logic/robotMovements/RobotMovements.cpp 	modified:   lib/logic/robotMovements/RobotMovements.h

+++ lib/components/servo/ServoMotor.h
@@ -19,8 +19,8 @@ class ServoMotor
         void moveToAngleFast(int angle);
         int getAngle() const { return _currentAngle; }
 
-        static const uint8_t MIN_ANGLE = 5;
-        static const uint8_t MAX_ANGLE = 175;
+        static const uint16_t MIN_ANGLE = 5;
+        static const uint16_t MAX_ANGLE = 175;
 };
 
 #endif
